+++ packages/at_secondary_server/lib/src/verb/handler/abstract_verb_handler.dart
@@ -1,5 +1,6 @@
 import 'dart:collection';
 import 'dart:convert';
+
 import 'package:at_commons/at_commons.dart';
 import 'package:at_persistence_secondary_server/at_persistence_secondary_server.dart';
 import 'package:at_secondary/src/connection/inbound/inbound_connection_metadata.dart';
@@ -7,12 +8,13 @@ import 'package:at_secondary/src/constants/enroll_constants.dart';
 import 'package:at_secondary/src/enroll/enroll_datastore_value.dart';
 import 'package:at_secondary/src/server/at_secondary_impl.dart';
 import 'package:at_secondary/src/utils/handler_util.dart' as handler_util;
+import 'package:at_secondary/src/verb/handler/otp_verb_handler.dart';
+import 'package:at_secondary/src/utils/secondary_util.dart';
 import 'package:at_secondary/src/verb/handler/sync_progressive_verb_handler.dart';
 import 'package:at_secondary/src/verb/manager/response_handler_manager.dart';
 import 'package:at_server_spec/at_server_spec.dart';
 import 'package:at_server_spec/at_verb_spec.dart';
 import 'package:at_utils/at_logger.dart';
-import 'package:at_secondary/src/utils/secondary_util.dart';
 
 final String paramFullCommandAsReceived = 'FullCommandAsReceived';
 
@@ -190,8 +192,13 @@ abstract class AbstractVerbHandler implements VerbHandler {
       }
     }
     // set passed namespace. If passed namespace is null, get namespace from atKey
-    final keyNamespace =
-        namespace ?? (atKey != null ? AtKey.fromString(atKey).namespace : null);
+    String? keyNamespace;
+    try {
+      keyNamespace = namespace ??
+          (atKey != null ? AtKey.fromString(atKey).namespace : null);
+    } catch (e) {
+      throw AtEnrollmentException('AtKey.toString($atKey) failed: $e');
+    }
     if (keyNamespace == null && atKey == null) {
       logger.shout('Both AtKey and namespace are null');
       return false;
@@ -263,30 +270,55 @@ abstract class AbstractVerbHandler implements VerbHandler {
   ///
   /// Additionally, this function removes the OTP from the keystore to prevent
   /// its reuse.
-  Future<bool> isOTPValid(String? otp) async {
-    if (otp == null) {
+  Future<bool> isPasscodeValid(String? passcode) async {
+    if (passcode == null) {
       return false;
     }
-    // Check if user have configured SPP(Semi-Permanent Pass-code).
+    // 1. Check if user has configured an SPP(Semi-Permanent Pass-code).
     // If SPP key is available, check if the otp sent is a valid pass code.
     // If yes, return true, else check it is a valid OTP.
-    String sppKey =
-        'private:spp${AtSecondaryServerImpl.getInstance().currentAtSign}';
-    if (keyStore.isKeyExists(sppKey)) {
-      AtData atData = await keyStore.get(sppKey);
-      if (atData.data?.toLowerCase() == otp.toLowerCase()) {
-        return true;
-      }
+    String passcodeKey = OtpVerbHandler.passcodeKey(passcode, isSpp: true);
+    if (!keyStore.isKeyExists(passcodeKey)) {
+      // if new SPPKey does not exist in keystore, check for SPP data against legacy SPP key
+      // New SPP key has __otp namespace, legacy key does NOT have any namespace
+      passcodeKey =
+          'private:spp${AtSecondaryServerImpl.getInstance().currentAtSign}';
     }
-    // If SPP is not valid, then check if the provided otp is valid.
-    String otpKey =
-        'private:${otp.toLowerCase()}${AtSecondaryServerImpl.getInstance().currentAtSign}';
-    AtData otpAtData;
     try {
-      otpAtData = await keyStore.get(otpKey);
+      AtData? sppAtData = await keyStore.get(passcodeKey);
+      // SPP has a special key so we have to check the value that was stored
+      // (which is the actual SPP)
+      // By comparison, OTPs are stored with the key being ${OTP}.__otp@alice
+      // i.e. the OTP is part of the key, and the stored data is irrelevant
+      if (sppAtData?.data?.toLowerCase() == passcode.toLowerCase()) {
+        if (SecondaryUtil.isActiveKey(sppAtData)) {
+          return true;
+        } else {
+          logger.finest(
+              'SPP found in KeyStore but has expired. Validating as OTP');
+        }
+      }
     } on KeyNotFoundException {
+      logger.finest('No SPP found in KeyStore. Validating as OTP');
+    }
+
+    // 2. If not a valid SPP, then check against OTP keys
+    String otpKey = OtpVerbHandler.passcodeKey(passcode, isSpp: false);
+    if (!keyStore.isKeyExists(otpKey)) {
+      // if new OTPKey does not exist in keystore, check for OTP data against legacy OTPKey
+      // New OTP key has __otp namespace, legacy key does not have namespace
+      otpKey =
+          'private:${passcode.toLowerCase()}${AtSecondaryServerImpl.getInstance().currentAtSign}';
+    }
+
+    AtData? otpAtData;
+    try {
+      otpAtData ??= await keyStore.get(otpKey);
+    } on KeyNotFoundException {
+      logger.finer('OTP NOT found in KeyStore');
       return false;
     }
+
     bool isOTPValid = SecondaryUtil.isActiveKey(otpAtData);
     // Remove the OTP after it is used.
     // NOTE: SPP code should NOT be deleted. only OTPs should be